+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/payment/controller/PaymentController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.NBE_4_5_2.Team5.domain.payment.dto.PaymentDto;
 import com.NBE_4_5_2.Team5.domain.payment.dto.PaymentMetaData;
 import com.NBE_4_5_2.Team5.domain.payment.service.PaymentService;
-import com.NBE_4_5_2.Team5.global.response.RsData;
+import com.NBE_4_5_2.Team5.global.dto.RsData;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
@@ -23,6 +23,14 @@ public class PaymentController {
 
 	private final PaymentService paymentService;
 
+	@GetMapping("/metadata")
+	public RsData<PaymentMetaData> saveMetaData(@NotNull @RequestParam String id,
+		@NotNull @RequestParam Integer amount) {
+		PaymentMetaData metadata = paymentService.saveMetaData(id, amount);
+		return new RsData<>("결제 메타데이터 저장 성공.", HttpStatus.OK.toString(), metadata);
+	}
+
+
 	@GetMapping("/request")
 	public RsData<Void> requestPayment(@RequestParam(name = "orderId") @NotNull String id,
 		@RequestParam(name = "paymentKey") @NotNull String paymentKey,
@@ -32,12 +40,6 @@ public class PaymentController {
 		return new RsData<>("페이 충전 결제 요청 성공. ", HttpStatus.OK.toString(), null);
 	}
 
-	@GetMapping("/metadata")
-	public RsData<PaymentMetaData> saveMetaData(@NotNull @RequestParam String id,
-		@NotNull @RequestParam Integer amount) {
-		PaymentMetaData metadata = paymentService.saveMetaData(id, amount);
-		return new RsData<>("결제 메타데이터 저장 성공.", HttpStatus.OK.toString(), metadata);
-	}
 
 	public record PurchaseItemReqDto(String productId) {
 	}
