+++ backend/src/main/java/com/NBE_4_5_2/Team5/global/aspect/ResponseAspect.java
@@ -5,7 +5,7 @@ import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
 import org.springframework.stereotype.Component;
 
-import com.NBE_4_5_2.Team5.global.response.RsData;
+import com.NBE_4_5_2.Team5.global.dto.RsData;
 
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
