--- backend/src/main/java/com/NBE_4_5_2/Team5/global/response/Empty.java
@@ -1,7 +1,0 @@
-package com.NBE_4_5_2.Team5.global.response;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class Empty {
-}
