--- backend/src/main/java/com/NBE_4_5_2/Team5/global/response/RsData.java
@@ -1,33 +1,0 @@
-package com.NBE_4_5_2.Team5.global.response;
-
-import org.springframework.lang.NonNull;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonInclude;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@Getter
-@AllArgsConstructor
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class RsData<T> {
-	@NonNull
-	private String code;
-	@NonNull
-	private String message;
-	@NonNull
-	private T data;
-
-	public RsData(String code, String message) {
-		this(code, message, (T)new Empty());
-	}
-
-	@JsonIgnore
-	public int getStatusCode() {
-		String statusCodeStr = code.split("-")[0];
-		return Integer.parseInt(statusCodeStr);
-	}
-}
